requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.2.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Django==3.2.8
django-autoslug==1.9.8
django-debug-toolbar==3.2.2
django-environ==0.8.0
django-extensions==3.1.3
django-js-asset==1.2.2
django-model-utils==4.2.0
django-modeltranslation==0.17.3
django-mptt==0.13.4
django-view-breadcrumbs==2.2.2
djangorestframework==3.12.4
Faker==9.5.2
flake8==4.0.1
flake8-isort==4.1.1
idna==3.3
isort==5.9.3
kombu==5.2.2
mccabe==0.6.1
pika==1.2.0
Pillow==8.4.0
prompt-toolkit==3.0.24
pycodestyle==2.8.0
pyflakes==2.4.0
python-dateutil==2.8.2
pytz==2021.3
redis==4.0.2
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
testfixtures==6.18.3
text-unidecode==1.3
tqdm==4.62.3
transliterate==1.10.2
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
psycopg2-binary==2.9.2
gunicorn==20.1.0
